+++ src/components/resource-edit-drawer.tsx
@@ -1,4 +1,3 @@
-import { Switch } from "@headlessui/react";
 import { type V1ObjectMeta } from "@kubernetes/client-node";
 import Editor, { DiffEditor } from "@monaco-editor/react";
 import yaml from "js-yaml";
@@ -7,42 +6,12 @@ import { useEffect, useState } from "react";
 import { useDefaultLanguage } from "../providers/default-language-provider";
 import { useTheme } from "../providers/theme-provider";
 import { Drawer } from "./drawer";
+import { ToggleButton } from "./toggle-button";
 
 interface ResourceEditDrawerProps<T> {
   isOpen: boolean;
   selectedResource: T | null;
   handleClose: () => void;
-}
-
-interface ToggleButtonProps {
-  checked: boolean;
-  onChange: (checked: boolean) => void;
-  checkedLabel: string;
-  uncheckedLabel: string;
-}
-function ToggleButton({ checked, onChange, checkedLabel, uncheckedLabel }: ToggleButtonProps) {
-  const label = checked ? checkedLabel : uncheckedLabel;
-  return (
-    <Switch.Group as="div" className="flex items-center">
-      <Switch
-        checked={checked}
-        onChange={onChange}
-        className={`${
-          checked ? "bg-blue-600 dark:bg-blue-300" : "bg-gray-200 dark:bg-gray-500"
-        } relative inline-flex h-6 w-11 items-center rounded-full`}
-      >
-        <span className="sr-only">{label}</span>
-        <span
-          className={`${
-            checked ? "translate-x-6" : "translate-x-1"
-          } inline-block h-4 w-4 rounded-full bg-white transition`}
-        />
-      </Switch>
-      <Switch.Label as="span" className="ml-2">
-        <span className="text-sm font-medium text-gray-900 dark:text-gray-100">{label}</span>
-      </Switch.Label>
-    </Switch.Group>
-  );
 }
 
 function getEditorTheme(
